centralise serialisation methods as well
