+++ Grasshopper_Engine/Objects/Types/GH_IObject.cs
@@ -33,7 +33,7 @@ using BH.Engine.Serialiser;
 
 namespace BH.Engine.Grasshopper.Objects
 {
-    public class IObjectGoo : GH_BHoMGoo<IObject>, IGH_PreviewData, GH_ISerializable
+    public class IObjectGoo : GH_BHoMGoo<IObject>, IGH_PreviewData
     {
         /*******************************************/
         /**** Properties                        ****/
@@ -93,28 +93,6 @@ namespace BH.Engine.Grasshopper.Objects
                 return "null";
             else
                 return val.ToString();
-        }
-
-        /***************************************************/
-
-        public override bool Read(GH_IReader reader)
-        {
-            string json = "";
-            reader.TryGetString("Json", ref json);
-
-            if (json != null && json.Length > 0)
-                Value = (IObject)BH.Engine.Serialiser.Convert.FromJson(json);
-
-            return true;
-        }
-
-        /***************************************************/
-
-        public override bool Write(GH_IWriter writer)
-        {
-            if (Value != null)
-                writer.SetString("Json", Value.ToJson());
-            return true;
         }
 
         /***************************************************/
