+++ Grasshopper_Engine/Objects/Types/GH_Type.cs
@@ -20,15 +20,13 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using BH.Engine.Serialiser;
 using GH_IO;
-using GH_IO.Serialization;
 using Grasshopper.Kernel.Types;
 using System;
 
 namespace BH.Engine.Grasshopper.Objects
 {
-    public class GH_Type : GH_BHoMGoo<Type>, GH_ISerializable
+    public class GH_Type : GH_BHoMGoo<Type>
     {
         /*******************************************/
         /**** Properties                        ****/
@@ -90,26 +88,5 @@ namespace BH.Engine.Grasshopper.Objects
 
         /***************************************************/
 
-        public override bool Read(GH_IReader reader)
-        {
-            string json = "";
-            reader.TryGetString("Json", ref json);
-
-            if (json != null && json.Length > 0)
-                Value = (Type)BH.Engine.Serialiser.Convert.FromJson(json);
-
-            return true;
-        }
-
-        /***************************************************/
-
-        public override bool Write(GH_IWriter writer)
-        {
-            if (Value != null)
-                writer.SetString("Json", Value.ToJson());
-            return true;
-        }
-
-        /***************************************************/
     }
 }
